basic styling with scss added, select to toggle view moved to new component, however still does not work
