+++ app/src/app.js
@@ -1,6 +1,7 @@
 import angular from 'angular';
 import components from './components';
 import services from './services';
+import './scss/main.scss';
 
 const app = angular.module('myApp', [
     components,
