+++ app/src/components/image-app/image-app.html
@@ -4,6 +4,10 @@
     <option value="thumbnail">thumbnail</option>
     <option value="full-size">full-size</option>
 </select>
+<!--<view-filter 
+    view = "app.view">
+<view-filter>-->
+
 
 <div ng-if="app.loading">Loading...</div>
