+++ app/src/components/image-app/image-app.js
@@ -1,5 +1,5 @@
 import template from './image-app.html';
-import styles from './image-app.css';
+import styles from './image-app.scss';
 export default {
     template,
     controller,
@@ -11,6 +11,12 @@ controller.$inject = ['imageService'];
 function controller(images) {
     this.loading = true;
     this.styles = styles;
+    this.view;
+    this.selectedView = 'description';
+
+    // $scope.$watch('app.view', () => {
+    //     this.selectedView = 'description';
+    // });
 
     images.get().then(images => {
         this.loading = false;
