+++ app/src/components/image-app/image-app.scss
@@ -1,0 +1,19 @@
+ul {
+    list-style-type: none;
+    width: 80%;
+    margin: 0 auto;
+}
+
+h2 {
+    text-align: center;
+    font-size: 2.5em;
+    margin: .5em 0;
+}
+
+.thumbnails li {
+    display: inline-block;
+}
+
+select {
+    margin: .5em 0 .1em 11%;
+}
