+++ app/src/components/image-description/image-description.html
@@ -1,5 +1,8 @@
 <div>
     <h3>{{$ctrl.image.title}}</h3> 
-    <a href="{{$ctrl.image.url}}">Link</a>
-    <p>Description: {{$ctrl.image.description}}</p>
+
+    <section>
+        <a href="{{$ctrl.image.url}}">Link</a>
+        <p>Description: {{$ctrl.image.description}}</p>
+    </section>
 <div>
