+++ app/src/components/image-description/image-description.js
@@ -1,9 +1,12 @@
 import template from './image-description.html';
+import styles from './image-description.scss';
 
 export default {
     template,
     bindings: {
         image: '='
     },
-    controller() {}
+    controller() {
+        this.styles = styles;
+    }
 };
