+++ app/src/components/image-description/image-description.scss
@@ -1,0 +1,3 @@
+section {
+    // margin-left: 10px;
+}
