+++ app/src/components/image-full/image-full.js
@@ -1,9 +1,12 @@
 import template from './image-full.html';
+import styles from './image-full.scss';
 
 export default {
     template,
     bindings: {
         image: '='
     },
-    controller() {}
+    controller() {
+        this.styles = styles;
+    }
 };
