+++ app/src/components/image-full/image-full.scss
@@ -1,0 +1,5 @@
+img {
+    max-width: 100%;
+    display: block;
+    margin: 0 auto;
+}
