+++ app/src/components/image-thumbnail/image-thumbnail.js
@@ -1,5 +1,5 @@
 import template from './image-thumbnail.html';
-import styles from './image-thumbnail.css';
+import styles from './image-thumbnail.scss';
 
 export default {
     template,
