+++ app/src/components/image-thumbnail/image-thumbnail.scss
@@ -1,5 +1,5 @@
 :local(.thumbnail) {
   height: 100px;
   width: 100px;
-  display: inline-block;  
 }
+
