+++ app/src/components/view-filter/view-filter.js
@@ -1,0 +1,16 @@
+import template from './view-filter.html';
+
+export default {
+    template,
+    bindings: {
+        view: '='
+    },
+    controller() {
+        this.view = '';
+
+        this.changeView = view => {
+            this.view = view;
+            console.log(this.view);
+        };
+    }
+};
