+++ app/src/index.html
@@ -4,7 +4,9 @@
         <title>Image Widget</title>
     </head>
     <body>
-        <header>Image Widget</header>
+        <header>
+            <h1>Image Widget</h1>
+        </header>
         <main>
             <image-app></image-app>
         </main>
