+++ app/src/scss/partials/_color.scss
@@ -1,0 +1,3 @@
+$primaryColor: #9C27B0;
+$primaryFont: #212121;
+$outlineColor: #BDBDBD;
