+++ app/webpack.config.js
@@ -13,7 +13,7 @@ module.exports = {
         new HtmlWebpackPlugin({
             template: './src/index.html'
         }),
-        new ExtractTextPlugin('styles.css')
+        // new ExtractTextPlugin('styles.css')
     ],
 
     module: {
@@ -30,11 +30,15 @@ module.exports = {
                 presets: ['es2015']
             }
         }, {
-            test: /\.css$/,
-            loader: ExtractTextPlugin.extract('style-loader', 'css-loader')
+            test: /\.scss$/,
+            // loader: ExtractTextPlugin.extract('style-loader', 'css-loader')
+            loader: 'style-loader!css-loader?sourceMap!sass-loader?sourceMap'
         }, {
             test: /\.html$/,
             loader: 'html-loader'
         }]
+    },
+    sassLoader: {
+        includePaths: ['./src/scss/partials']
     }
 };
