+++ server/public/index.html
@@ -2,9 +2,11 @@
 <html ng-app="myApp">
     <head>
         <title>Image Widget</title>
-    <link href="styles.css" rel="stylesheet"></head>
+    </head>
     <body>
-        <header>Image Widget</header>
+        <header>
+            <h1>Image Widget</h1>
+        </header>
         <main>
             <image-app></image-app>
         </main>
